requirements.txt
flask==3.0.2
mistralai==0.1.3
python-dotenv==1.0.1 